*** PACKAGE CREATION FOR RS2014; * But missing [SNG Writer, Showlights.xml "WORKING" and Manifest.json/hsan "WORKING, ALMOST DONE"; * Fixed audio preview not selected; * Fixed game version not loaded/saved; * Fixed MAC platform no packing / Soundbank not generating for MAC; * Minor fixes.

+++ RocksmithToolkitLib/XML/Song.cs
@@ -506,7 +506,7 @@ namespace RocksmithToolkitLib.Xml
         public Single Time { get; set; }
 
         [XmlAttribute("measure")]
-        public Int32 Measure { get; set; }
+        public Int16 Measure { get; set; }
     }
 
     [XmlType("section")]
